+++ genDataGetProcess.h
@@ -17,15 +17,15 @@ typedef enum
 	YESGPS
 }gpsflag_t;
 
-extern double zI[10];
-extern double zG[7];
-extern double gpstime;
-extern double head;
-extern double speed;
+extern float zI[10];
+extern float zG[7];
+extern float gpstime;
+extern float head;
+extern float speed;
 extern gpsflag_t gpsflag;
 static _Bool ISstarted;
 //extern uint8_t gpsbuff[RXBUFF_SIZE];
-void pos2googAddr(double* dest, uint8_t* from, uint8_t length, uint8_t _cPos);
-void ToDoubleAddr(double* dest, uint8_t* from, uint8_t length);
+void pos2googAddr(float* dest, uint8_t* from, uint8_t length, uint8_t _cPos);
+void ToDoubleAddr(float* dest, uint8_t* from, uint8_t length);
 void GPSDataProcess(void);
 void INSDataProcess(void);
