+++ result/insgps_v4_0.h
@@ -17,10 +17,10 @@
 #include "Cbn_31_types.h"
 
 /* Function Declarations */
-void insgps_v4_0(const double zI[10], const double zG[7], double gpsflag, double
-                 dt, double g0, double a, double e, double we, const double Q
-                 [144], const double R[36], double PVA[10], double bias[6],
-                 double Pk_1[225], double xk_1[15]);
+void insgps_v4_0(const float zI[10], const float zG[7], int gpsflag, float
+                 dt, float g0, float a, float e, float we, const float Q
+                 [144], const float R[36], float PVA[10], float bias[6],
+                 float Pk_1[225], float xk_1[15]);
 
 #endif
